Fix/pending jobs invisible if q specified (#4)
* Removed filter paramter.
When using load_values as a requestable paramaeter in job (e.g. -l mem_free=2G). It will not be shown in case the -q param is specified in qstat.

This is due to the way how load_values are used. Scheduler will assign the queue at the moment of start of the job, as there is no way for scheduler to know, how the load_value will behave in the future. (kinda like Schroedinger job...)

---------

Co-authored-by: Jan Pesan <janp@speechmatics.com>